requirements-win.txt
appdirs==1.4.4
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
distlib==0.3.9
docutils==0.19
filelock==3.16.1
idna==3.4
Jinja2==3.1.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.0.2
MarkupSafe==2.1.1
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.1
pep517==0.13.0
pexpect==4.8.0
Pillow==9.5.0
platformdirs==4.3.6
plyer==2.0.0
ptyprocess==0.7.0
pycryptodomex==3.20.0
Pygments==2.13.0
pyparsing==3.0.9
pypiwin32==223
pypng==0.20220715.0
pywin32==305
pyzbar==0.1.9
qrcode==7.4.2
requests==2.28.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.12
virtualenv==20.28.1
zipp==3.8.1
